Play nicer with NearFuture
Using `+` rather than `@` creates a new part, rather than editing the existing part. This seems to be the desire, as a new name is provided.

The issue I ran into was I'd created a microsat with a battery, and then installed NearFutureElectrical, and KSP deleted my microsat vessel on loading the game. The above change fixed that.

I haven't tested the Argon Container fix, but I expect it to work the same way.